Stream 5.2 Библиотека Spring-JDBC - работа с ключами, транзакции, адаптеры

+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/web/JdbcSpringTest.java
@@ -1,11 +1,13 @@
 package guru.qa.niffler.test.web;
 
 import guru.qa.niffler.data.entity.auth.AuthUserEntity;
-import guru.qa.niffler.model.*;
-import guru.qa.niffler.service.auth.AuthUserDBClient;
-import guru.qa.niffler.service.userdata.UserdataDBClient;
+import guru.qa.niffler.model.CategoryJson;
+import guru.qa.niffler.model.SpendJson;
+import guru.qa.niffler.model.UserdataUserJson;
 import guru.qa.niffler.service.spend.CategoryDBSpringClient;
-import guru.qa.niffler.service.spend.SpendDBClient;
+import guru.qa.niffler.service.spend.SpendDBSpringClient;
+import guru.qa.niffler.service.auth.AuthUserDBSpringClient;
+import guru.qa.niffler.service.userdata.UserdataDBSpringClient;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -14,22 +16,23 @@ import java.util.Date;
 
 import static guru.qa.niffler.config.Constants.MAIN_USERNAME;
 import static guru.qa.niffler.model.CurrencyValues.RUB;
-import static guru.qa.niffler.util.RandomDataUtils.*;
+import static guru.qa.niffler.util.RandomDataUtils.randomName;
+import static guru.qa.niffler.util.RandomDataUtils.randomSurname;
 
 @Slf4j
-class JdbcTest {
+class JdbcSpringTest {
 
     @Test
-    void daoCRUDSpendTest() {
+    void daoCRUDSpringSpendTest() {
         String categoryName = "top cource";
         String spendDescription = "cool cource";
-        SpendDBClient dbSpend = new SpendDBClient();
-        dbSpend.findByUsernameAndDescription(MAIN_USERNAME, spendDescription)
-                .ifPresent(entity -> {
-                    dbSpend.delete(entity);
-                    log.info("SPEND DELETED");
-                });
-        SpendJson newSpendJson = dbSpend.create(new SpendJson(
+        SpendDBSpringClient dbSpendSpring = new SpendDBSpringClient();
+        SpendJson spendJson = dbSpendSpring.findByUsernameAndDescription(MAIN_USERNAME, spendDescription);
+        if (spendJson != null) {
+            dbSpendSpring.delete(spendJson);
+            log.info("SPEND DELETED");
+        }
+        SpendJson newSpendJson = dbSpendSpring.create(new SpendJson(
                 null,
                 new Date(),
                 new CategoryJson(null,
@@ -44,8 +47,8 @@ class JdbcTest {
         log.info("SPEND CREATED");
         log.info("CATEGORY CREATED");
         Assertions.assertNotNull(newSpendJson, "Объект Spend не создан");
-        Assertions.assertNotNull(newSpendJson.category(), "Объект Category не создан");
-        SpendJson resultSpend = dbSpend.findByUsernameAndDescription(MAIN_USERNAME, spendDescription).get();
+        Assertions.assertNotNull(newSpendJson.category(),"Объект Category не создан");
+        SpendJson resultSpend = dbSpendSpring.findByUsernameAndDescription(MAIN_USERNAME, spendDescription);
         Assertions.assertTrue(newSpendJson.username().equals(resultSpend.username())
                         && newSpendJson.id().equals(resultSpend.id())
                         && newSpendJson.category().id().equals(resultSpend.category().id()),
@@ -55,7 +58,7 @@ class JdbcTest {
     }
 
     @Test
-    void checkCRUDCategoryTest() {
+    void checkCRUDCategorySpringTest() {
         String categoryName = "Top category";
         CategoryDBSpringClient dbSpringClient = new CategoryDBSpringClient();
         dbSpringClient.findByUsernameAndName(MAIN_USERNAME, categoryName)
@@ -78,14 +81,14 @@ class JdbcTest {
 
     @Test
     void daoCreateAndCheckUserTest() {
-        UserdataDBClient userdataDBClient = new UserdataDBClient();
-        AuthUserDBClient authUserDBClient = new AuthUserDBClient();
-        String username = "twix";
-        userdataDBClient.findByUsername(username)
-                .ifPresent(userJson -> {
-                    userdataDBClient.delete(userJson);
-                    log.info("DELETED USER");
-                });
+        UserdataDBSpringClient userdataDBClient = new UserdataDBSpringClient();
+        AuthUserDBSpringClient authUserDBClient = new AuthUserDBSpringClient();
+        String username = "twixSpring";
+        UserdataUserJson userJson = userdataDBClient.findByUsername(username);
+        if (userJson != null) {
+            userdataDBClient.delete(userJson);
+            log.info("DELETED USER");
+        }
         log.info("CREATE USER");
         UserdataUserJson userdataUserJson = userdataDBClient.create(
                 new UserdataUserJson(
@@ -99,36 +102,36 @@ class JdbcTest {
                         null,
                         null));
         log.info("CHECK USER");
-        UserdataUserJson userJson = userdataDBClient.findById(userdataUserJson.id()).get();
+        userJson = userdataDBClient.findById(userdataUserJson.id());
         Assertions.assertNotNull(userdataUserJson, "Юзер %s не создан".formatted(userdataUserJson.toString()));
         Assertions.assertTrue(userdataUserJson.id().equals(userJson.id())
                 && userdataUserJson.username().equals(userJson.username()), "Юзер %s не создан, а создан %s"
                 .formatted(userJson.toString(), userdataUserJson.toString()));
-        userJson = userdataDBClient.findByUsername(username).get();
+        userJson = userdataDBClient.findByUsername(username);
         Assertions.assertEquals(userdataUserJson.id(), userJson.id(), "Юзер %s не создан, а создан %s"
                 .formatted(userJson.toString(), userdataUserJson.toString()));
         log.info("CHECK AUTH_USER");
-        AuthUserEntity authUserEntity = authUserDBClient.findUserByUsername(userJson.username()).get();
+        AuthUserEntity authUserEntity = authUserDBClient.findUserByUsername(userJson.username());
         Assertions.assertNotNull(authUserEntity, "Юзер авторизации %s не был создан".formatted(authUserEntity));
-        Assertions.assertEquals(2, authUserDBClient.findAuthorityByUserId(authUserEntity).size(), "Authority не был создан");
+        Assertions.assertFalse(authUserDBClient.findAuthorityByUserId(authUserEntity).isEmpty(), "Authority не был создан");
         log.info("DELETE USER");
         userdataDBClient.delete(userdataUserJson);
-        userJson = userdataDBClient.findById(userdataUserJson.id()).orElse(null);
+        userJson = userdataDBClient.findById(userdataUserJson.id());
         Assertions.assertNull(userJson, "Юзер %s не удален".formatted(userJson));
-        authUserEntity = authUserDBClient.findUserByUsername(userdataUserJson.username()).orElse(null);
+        authUserEntity = authUserDBClient.findUserByUsername(userdataUserJson.username());
         Assertions.assertNull(authUserEntity, "Юзер авторизации %s был создан".formatted(authUserEntity));
     }
 
     @Test
     void checkTransactionByUserTest() {
-        UserdataDBClient userdataDBClient = new UserdataDBClient();
-        AuthUserDBClient authUserDBClient = new AuthUserDBClient();
+        UserdataDBSpringClient userdataDBClient = new UserdataDBSpringClient();
+        AuthUserDBSpringClient authUserDBClient = new AuthUserDBSpringClient();
         String username = "transac";
-        userdataDBClient.findByUsername(username)
-                .ifPresent(userJson -> {
-                    userdataDBClient.delete(userJson);
-                    log.info("DELETED USER");
-                });
+        UserdataUserJson userJson = userdataDBClient.findByUsername(username);
+        if (userJson != null) {
+            userdataDBClient.delete(userJson);
+            log.info("DELETED USER");
+        }
         Assertions.assertThrows(RuntimeException.class, () ->
                 userdataDBClient.create(
                         new UserdataUserJson(
@@ -143,10 +146,10 @@ class JdbcTest {
                                 null
                         )));
         log.info("EXPECT NOT CREATED USER");
-        UserdataUserJson udUser = userdataDBClient.findByUsername(username).orElse(null);
+        UserdataUserJson udUser = userdataDBClient.findByUsername(username);
         Assertions.assertNull(udUser, "Юзер %s был создан".formatted(udUser));
         log.info("EXPECT NOT CREATED AUTH_USER");
-        AuthUserEntity authUserEntity = authUserDBClient.findUserByUsername(username).orElse(null);
+        AuthUserEntity authUserEntity = authUserDBClient.findUserByUsername(username);
         Assertions.assertNull(authUserEntity, "Юзер авторизации %s был создан".formatted(authUserEntity));
     }
 }
